// lib/supabase/types.ts
export type Database = {
// ...
  public: {
    Tables: {
// ...
      patients: {
        Row: {
          care_level_medicine: string | null
          care_level_nursing: number | null
          created_at: string
          id: string
          location_bed: number | null
          location_room: string | null
          name: string
          note: string | null
          personal_number: string | null
          planned_check_in: string | null
          planned_check_out: string | null
          planned_operation: string | null
          quick_icons: string[]
          updated_at: string
        }
        Insert: {
          care_level_medicine?: string | null
          care_level_nursing?: number | null
          created_at?: string
          id?: string
          location_bed?: number | null
          location_room?: string | null
          name: string
          note?: string | null
          personal_number?: string | null
          planned_check_in?: string | null
          planned_check_out?: string | null
          planned_operation?: string | null
          quick_icons?: string[]
          updated_at?: string
        }
        Update: {
          care_level_medicine?: string | null
          care_level_nursing?: number | null
          created_at?: string
          id?: string
          location_bed?: number | null
          location_room?: string | null
          name?: string
          note?: string | null
          personal_number?: string | null
          planned_check_in?: string | null
          planned_check_out?: string | null
          planned_operation?: string | null
          quick_icons?: string[]
          updated_at?: string
        }
        Relationships: []
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}

type DatabaseWithoutInternals = Omit<Database, "__InternalSupabase">

type DefaultSchema = DatabaseWithoutInternals[Extract<keyof Database, "public">]
// ...
export type TablesUpdate<
  DefaultSchemaTableNameOrOptions extends
    | keyof DefaultSchema["Tables"]
    | { schema: keyof DatabaseWithoutInternals },
  TableName extends DefaultSchemaTableNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaTableNameOrOptions["schema"]]["Tables"]
    : never = never,
> = DefaultSchemaTableNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[DefaultSchemaTableNameOrOptions["schema"]]["Tables"][TableName] extends {
      Update: infer U
    }
    ? U
    : never
  : DefaultSchemaTableNameOrOptions extends keyof DefaultSchema["Tables"]
    ? DefaultSchema["Tables"][DefaultSchemaTableNameOrOptions] extends {
        Update: infer U
      }
      ? U
      : never
    : never